+++ src/main/java/com/example/demo/store/service/impl/AdminStoreServiceImpl.java
@@ -4,18 +4,31 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import com.example.demo.store.repository.ProductCategoryRepository;
+import com.example.demo.store.repository.ProductImgRepository;
 import com.example.demo.store.repository.ProductRepository;
 import com.example.demo.store.repository.ProductStockRepository;
 import com.example.demo.store.service.AdminStoreService;
-import com.example.demo.store.service.ProductService;
+
+import jakarta.transaction.Transactional;
+
+import com.amazonaws.services.kms.model.NotFoundException;
+import com.example.demo.ncp.dto.NCPObjectStorageDTO;
+import com.example.demo.ncp.service.NCPObjectStorageService;
 import com.example.demo.store.dto.AdminProductDTO;
+import com.example.demo.store.dto.ProductDTO;
+import com.example.demo.store.dto.ProductStockDTO;
+import com.example.demo.store.entity.ProductCategoryEntity;
 import com.example.demo.store.entity.ProductEntity;
+import com.example.demo.store.entity.ProductImgEntity;
 import com.example.demo.store.entity.ProductStockEntity;
 import com.example.demo.store.entity.ProductStockEntity.Size;
 
@@ -30,7 +43,13 @@ public class AdminStoreServiceImpl implements AdminStoreService {
     private ProductStockRepository stockRepository;
 
     @Autowired
-    private ProductService productService;
+    private ProductImgRepository imgRepository;
+
+    @Autowired
+    private ProductCategoryRepository categoryRepository;
+
+    @Autowired
+    private NCPObjectStorageService ncpObjectStorageService;
 
     // 24.12.13 - 상품 현황 조회 - yjy
     @Override
@@ -57,6 +76,175 @@ public class AdminStoreServiceImpl implements AdminStoreService {
         return result;
     }
 
+    // 상품 등록 - uj
+    @Override
+    @SuppressWarnings("unchecked")
+    public void saveProduct(Map<String, Object> map) {
+
+        // 1. 상품 DTO => Entity
+        ProductEntity productEntity = ProductEntity.toSaveProductEntity((ProductDTO) map.get("ProductDTO"));
+        // 1-1. 카테고리 Entity 가져와서 저장
+        ProductCategoryEntity categoryEntity = categoryRepository
+                .findById(Integer.parseInt(map.get("categoryId").toString())).orElse(null);
+        productEntity.setCategoryEntity(categoryEntity);
+
+        // 1-2. 상품 테이블에 저장
+        int productId = productRepository.save(productEntity).getId(); // DB에 저장된 기본키 반환
+        System.out.println("saveProduct: " + productId);
+
+        // 2. NCP & DB에 이미지 저장
+        for (int i = 1; i < 4; i++) {
+            List<MultipartFile> files = (List<MultipartFile>) map.get("layer" + i);
+            saveImgFiles(files, productEntity, i);
+        }
+
+        // 3. 재고 저장
+        // 사이즈 S, M, L
+        for (String size : new String[] { "S", "M", "L" }) {
+            ProductStockDTO stockDTO = new ProductStockDTO();
+            stockDTO.setSize(size);
+            stockDTO.setCount(Integer.parseInt(map.get(size).toString()));
+            ProductStockEntity stockEntity = ProductStockEntity.toSaveStockEntity(stockDTO, productEntity);
+
+            stockRepository.save(stockEntity);
+        }
+
+    }
+
+    // NCP 업로드 & 이미지 테이블에 Data 저장 (메인 Img ProductEntity에 설정) - uj
+    private void saveImgFiles(List<MultipartFile> files, ProductEntity product, int layer) {
+        for (MultipartFile file : files) {
+            try {
+                NCPObjectStorageDTO ncpDTO = new NCPObjectStorageDTO();
+                // NCP Object Storage 업로드
+                String fileName = ncpObjectStorageService.uploadFile(
+                        ncpDTO.getBUCKETNAME(), ncpDTO.getPRODUCTDIRECTORY(), file);
+
+                // ProductEntity에 메인 이미지 저장
+                if (layer == 1) {
+                    product.setImg(fileName);
+                    productRepository.save(product);
+                }
+
+                // 데이터베이스에 저장할 ProductImgEntity 생성
+                ProductImgEntity imgEntity = new ProductImgEntity();
+                imgEntity.setProductEntity(product);
+                imgEntity.setFileName(fileName);
+                imgEntity.setOriginalFileName(file.getOriginalFilename());
+                imgEntity.setLayer(layer); // 레이어 값 설정
+
+                // ProductImgRepository에 저장
+                imgRepository.save(imgEntity);
+
+                // 로그 출력
+                System.out.println("Saved image for layer " + layer + ": " + fileName);
+            } catch (Exception e) {
+                System.err.println("File upload failed for layer " + layer + ": " + e.getMessage());
+            }
+        }
+    }
+
+    // 상품 수정 - 24.11.25, 26, 27 - uj
+    @Override
+    @Transactional
+    @SuppressWarnings("unchecked")
+    public void putProduct(Map<String, Object> map) {
+        NCPObjectStorageDTO ncpDTO = new NCPObjectStorageDTO();
+
+        // 1. DB에서 상품 조회
+        ProductDTO productDTO = (ProductDTO) map.get("ProductDTO");
+        ProductEntity productEntity = productRepository.findById(productDTO.getId()).orElseThrow(
+                () -> new NotFoundException("해당 상품(" + productDTO.getId() + ")이 조회되지 않습니다."));
+
+        // 1-1. 사용자 입력 데이터
+        productEntity.setName(productDTO.getName());
+        productEntity.setContent(productDTO.getContent());
+        productEntity.setPrice(productDTO.getPrice());
+        productEntity.setGender(productDTO.getGender());
+
+        // 2. 카테고리 Entity 추출
+        ProductCategoryEntity categoryEntity = categoryRepository
+                .findById(Integer.parseInt(map.get("categoryId").toString()))
+                .orElseThrow(
+                        () -> new NotFoundException("해당 카테고리(" + map.get("categoryId").toString() + ")가 조회되지 않습니다."));
+        productEntity.setCategoryEntity(categoryEntity);
+
+        // 3. 이미지(NCP) 수정
+        // 이미지 Id
+        String[] selectImgId = (String[]) map.get("selectImgId"); // 존재할 상품 이미지 Id
+        List<ProductImgEntity> imgEntityList = productEntity.getImgList(); // 상품 이미지 Entity
+        List<ProductImgEntity> selectimgEntityList = new ArrayList<>(); // DB에 저장할 Img Entity
+
+        // 3-1. 이미지 삭제 (NCP & DB)
+        for (ProductImgEntity imgEntity : imgEntityList) {
+            boolean isDelete = true;
+
+            for (String imgIdStr : selectImgId) {
+                int imgId = Integer.parseInt(imgIdStr);
+                // DB에 남길 Img Entity
+                if (imgId == imgEntity.getId()) {
+                    selectimgEntityList.add(imgEntity);
+                    isDelete = false;
+                    break;
+                }
+            }
+
+            // NCP에서 삭제
+            if (isDelete) {
+                ncpObjectStorageService.deleteFile(
+                        ncpDTO.getBUCKETNAME(),
+                        ncpDTO.getPRODUCTDIRECTORY(),
+                        imgEntity.getFileName());
+
+                System.out.println("삭제 Img Entity Id: " + imgEntity.getId());
+            }
+        }
+
+        // DB 에서 삭제
+        productEntity.getImgList().clear(); // 기존 컬렉션 초기화
+        productEntity.getImgList().addAll(selectimgEntityList); // 새 값 추가
+
+        // 3-2. 이미지 업로드 (NCP & DB)
+        for (int i = 1; i < 5; i++) {
+            if (map.get("layer" + i) != null) {
+                List<MultipartFile> files = (List<MultipartFile>) map.get("layer" + i);
+                saveImgFiles(files, productEntity, i);
+            }
+        }
+
+        // 4. 재고 수정
+        int soldoutCount = 0;
+        List<ProductStockEntity> stockEntiyList = productEntity.getStockList();
+        for (ProductStockEntity stockEntity : stockEntiyList) {
+            switch (stockEntity.getSize()) {
+                case S:
+                    stockEntity.setCount(Integer.parseInt(map.get("S").toString()));
+                    break;
+                case M:
+                    stockEntity.setCount(Integer.parseInt(map.get("M").toString()));
+                    break;
+                case L:
+                    stockEntity.setCount(Integer.parseInt(map.get("L").toString()));
+                    break;
+                default:
+                    break;
+            }
+
+            soldoutCount += stockEntity.getCount() == 0 ? 1 : 0; // 사이즈 별 재고 0인것 개수 구하기
+        }
+
+        // 4-1. 재고 상태 설정
+        if (soldoutCount == 0) { // 모든 사이즈 재고: 0 이상
+            productEntity.setStatus(ProductEntity.ProductStatus.EXIST);
+        } else if (soldoutCount < stockEntiyList.size()) { // 일부 사이즈 재고 : 0
+            productEntity.setStatus(ProductEntity.ProductStatus.SOMESOLDOUT);
+        } else if (soldoutCount == stockEntiyList.size()) { // 모든 사이즈 재고 : 0
+            productEntity.setStatus(ProductEntity.ProductStatus.SOLDOUT);
+        }
+
+        productRepository.save(productEntity);
+    }
+
     // 관리자 상품목록 가져오기, 카테고리 or 키워드별 검색 후 페이징처리 - 12/16 17:31 tang
     // 24.12.18 - 반환 데이터 수정 - uj
     @Override
@@ -68,12 +256,6 @@ public class AdminStoreServiceImpl implements AdminStoreService {
         Pageable pageable = (Pageable) dataMap.get("pageable"); // 페이지 처리
         int block = Integer.parseInt(dataMap.get("block").toString()); // 한 페이지당 보여줄 숫자
         int sortby = Integer.parseInt(dataMap.get("sortby").toString()); // 정렬 기준 현재 없음
-        int categoryid = Integer.parseInt(dataMap.get("categoryid").toString());
-
-        String keyword = null; // 검색어
-        if (dataMap.get("keyword") != null) {
-            keyword = dataMap.get("keyword").toString();
-        }
 
         // 조건별 상품 목록 가져오기
         Page<ProductEntity> pageProductList = null;
@@ -93,9 +275,6 @@ public class AdminStoreServiceImpl implements AdminStoreService {
 
         // Entity -> DTO & 상품 재고 추출
         for (ProductEntity entity : pageProductList) {
-            // List<ProductStockEntity> stockEntityList =
-            // stockRepository.findByProductEntity(entity);
-
             // 사이즈에 따른 재고 추출
             Map<Size, Integer> stockMap = new HashMap<>();
             for (ProductStockEntity stock : entity.getStockList()) {
@@ -122,5 +301,41 @@ public class AdminStoreServiceImpl implements AdminStoreService {
         // products.put("category", productService.getCategory());
 
         return products;
+    }
+
+    // 24.12.11 - 상품 삭제 - sumin
+    @Override
+    public void deleteProduct(int productId) {
+        // productId에 해당하는 상품 조회
+        ProductEntity product = productRepository.findById(productId)
+                .orElseThrow(() -> new RuntimeException("해당 상품이 존재하지 않습니다."));
+        System.out.println("상품 번호: " + productId);
+        // delete 변수 true로 설정
+        product.setDelete(true);
+        product.setStatus(ProductEntity.ProductStatus.DELETED);
+
+        // 변경 사항 저장
+        productRepository.save(product);
+
+        System.out.println("상품 삭제 완료 상품 ID: " + productId);
+    }
+
+    // 24.12.11 - 상품 복구 - sumin
+    @Override
+    public boolean restoreProduct(int productId) {
+        Optional<ProductEntity> productEntity = productRepository.findById(productId);
+
+        if (productEntity.isPresent()) {
+            ProductEntity product = productEntity.get();
+
+            // 이미 삭제 상태인 경우만 복구
+            if (product.getDelete()) {
+                product.setDelete(false);
+                productRepository.save(product);
+                return true; // 복구 성공
+            }
+        }
+
+        return false; // 복구 실패
     }
 }
