+++ src/main/java/com/example/demo/store/service/impl/ProductServiceImpl.java
@@ -5,7 +5,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import com.example.demo.store.dto.*;
@@ -15,11 +14,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import com.amazonaws.services.kms.model.NotFoundException;
-import com.example.demo.ncp.dto.NCPObjectStorageDTO;
-import com.example.demo.ncp.service.NCPObjectStorageService;
 import com.example.demo.store.entity.ProductCategoryEntity;
 import com.example.demo.store.entity.ProductEntity;
 import com.example.demo.store.entity.ProductImgEntity;
@@ -27,12 +22,8 @@ import com.example.demo.store.entity.ProductStockEntity;
 import com.example.demo.store.entity.ReviewEntity;
 import com.example.demo.store.entity.ProductEntity.Gender;
 import com.example.demo.store.repository.ProductCategoryRepository;
-import com.example.demo.store.repository.ProductImgRepository;
 import com.example.demo.store.repository.ProductRepository;
-import com.example.demo.store.repository.ProductStockRepository;
 import com.example.demo.store.service.ProductService;
-
-import jakarta.transaction.Transactional;
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -40,16 +31,7 @@ public class ProductServiceImpl implements ProductService {
     private ProductRepository productRepository;
 
     @Autowired
-    private ProductImgRepository imgRepository;
-
-    @Autowired
     private ProductCategoryRepository categoryRepository;
-
-    @Autowired
-    private ProductStockRepository stockRepository;
-
-    @Autowired
-    private NCPObjectStorageService ncpObjectStorageService;
 
     // 상품 상세 정보 출력 - uj
     @Override
@@ -87,74 +69,6 @@ public class ProductServiceImpl implements ProductService {
         map.put("stockList", stockList);
 
         return map;
-    }
-
-    // 상품 등록 - uj
-    @Override
-    @SuppressWarnings("unchecked")
-    public void saveProduct(Map<String, Object> map) {
-
-        // 1. 상품 DTO => Entity
-        ProductEntity productEntity = ProductEntity.toSaveProductEntity((ProductDTO) map.get("ProductDTO"));
-        // 1-1. 카테고리 Entity 가져와서 저장
-        ProductCategoryEntity categoryEntity = categoryRepository
-                .findById(Integer.parseInt(map.get("categoryId").toString())).orElse(null);
-        productEntity.setCategoryEntity(categoryEntity);
-
-        // 1-2. 상품 테이블에 저장
-        int productId = productRepository.save(productEntity).getId(); // DB에 저장된 기본키 반환
-        System.out.println("saveProduct: " + productId);
-
-        // 2. NCP & DB에 이미지 저장
-        for (int i = 1; i < 4; i++) {
-            List<MultipartFile> files = (List<MultipartFile>) map.get("layer" + i);
-            saveImgFiles(files, productEntity, i);
-        }
-
-        // 3. 재고 저장
-        // 사이즈 S, M, L
-        for (String size : new String[] { "S", "M", "L" }) {
-            ProductStockDTO stockDTO = new ProductStockDTO();
-            stockDTO.setSize(size);
-            stockDTO.setCount(Integer.parseInt(map.get(size).toString()));
-            ProductStockEntity stockEntity = ProductStockEntity.toSaveStockEntity(stockDTO, productEntity);
-
-            stockRepository.save(stockEntity);
-        }
-
-    }
-
-    // NCP 업로드 & 이미지 테이블에 Data 저장 (메인 Img ProductEntity에 설정) - uj
-    private void saveImgFiles(List<MultipartFile> files, ProductEntity product, int layer) {
-        for (MultipartFile file : files) {
-            try {
-                NCPObjectStorageDTO ncpDTO = new NCPObjectStorageDTO();
-                // NCP Object Storage 업로드
-                String fileName = ncpObjectStorageService.uploadFile(
-                        ncpDTO.getBUCKETNAME(), ncpDTO.getPRODUCTDIRECTORY(), file);
-
-                // ProductEntity에 메인 이미지 저장
-                if (layer == 1) {
-                    product.setImg(fileName);
-                    productRepository.save(product);
-                }
-
-                // 데이터베이스에 저장할 ProductImgEntity 생성
-                ProductImgEntity imgEntity = new ProductImgEntity();
-                imgEntity.setProductEntity(product);
-                imgEntity.setFileName(fileName);
-                imgEntity.setOriginalFileName(file.getOriginalFilename());
-                imgEntity.setLayer(layer); // 레이어 값 설정
-
-                // ProductImgRepository에 저장
-                imgRepository.save(imgEntity);
-
-                // 로그 출력
-                System.out.println("Saved image for layer " + layer + ": " + fileName);
-            } catch (Exception e) {
-                System.err.println("File upload failed for layer " + layer + ": " + e.getMessage());
-            }
-        }
     }
 
     // 상품 리스트, 카테고리별 검색 or 키워드별 검색 후 페이징처리-11/25-tang & 11/27 - 페이징 처리 수정 - uj
@@ -244,94 +158,6 @@ public class ProductServiceImpl implements ProductService {
         return list;
     }
 
-    // 상품 수정 - 24.11.25, 26, 27 - uj
-    @Override
-    @Transactional
-    @SuppressWarnings("unchecked")
-    public void putProduct(Map<String, Object> map) {
-        NCPObjectStorageDTO ncpDTO = new NCPObjectStorageDTO();
-
-        // 1. DB에서 상품 조회
-        ProductDTO productDTO = (ProductDTO) map.get("ProductDTO");
-        ProductEntity productEntity = productRepository.findById(productDTO.getId()).orElseThrow(
-                () -> new NotFoundException("해당 상품(" + productDTO.getId() + ")이 조회되지 않습니다."));
-
-        // 1-1. 사용자 입력 데이터
-        productEntity.setName(productDTO.getName());
-        productEntity.setContent(productDTO.getContent());
-        productEntity.setPrice(productDTO.getPrice());
-        productEntity.setGender(productDTO.getGender());
-
-        // 2. 카테고리 Entity 추출
-        ProductCategoryEntity categoryEntity = categoryRepository
-                .findById(Integer.parseInt(map.get("categoryId").toString()))
-                .orElseThrow(
-                        () -> new NotFoundException("해당 카테고리(" + map.get("categoryId").toString() + ")가 조회되지 않습니다."));
-        productEntity.setCategoryEntity(categoryEntity);
-
-        // 3. 이미지(NCP) 수정
-        // 이미지 Id
-        String[] selectImgId = (String[]) map.get("selectImgId"); // 존재할 상품 이미지 Id
-        List<ProductImgEntity> imgEntityList = productEntity.getImgList(); // 상품 이미지 Entity
-        List<ProductImgEntity> selectimgEntityList = new ArrayList<>(); // DB에 저장할 Img Entity
-
-        // 3-1. 이미지 삭제 (NCP & DB)
-        for (ProductImgEntity imgEntity : imgEntityList) {
-            boolean isDelete = true;
-
-            for (String imgIdStr : selectImgId) {
-                int imgId = Integer.parseInt(imgIdStr);
-                // DB에 남길 Img Entity
-                if (imgId == imgEntity.getId()) {
-                    selectimgEntityList.add(imgEntity);
-                    isDelete = false;
-                    break;
-                }
-            }
-
-            // NCP에서 삭제
-            if (isDelete) {
-                ncpObjectStorageService.deleteFile(
-                        ncpDTO.getBUCKETNAME(),
-                        ncpDTO.getPRODUCTDIRECTORY(),
-                        imgEntity.getFileName());
-
-                System.out.println("삭제 Img Entity Id: " + imgEntity.getId());
-            }
-        }
-
-        // DB 에서 삭제
-        productEntity.getImgList().clear(); // 기존 컬렉션 초기화
-        productEntity.getImgList().addAll(selectimgEntityList); // 새 값 추가
-
-        // 3-2. 이미지 업로드 (NCP & DB)
-        for (int i = 1; i < 5; i++) {
-            if (map.get("layer" + i) != null) {
-                List<MultipartFile> files = (List<MultipartFile>) map.get("layer" + i);
-                saveImgFiles(files, productEntity, i);
-            }
-        }
-
-        // 4. 재고 수정
-        for (ProductStockEntity stockEntity : productEntity.getStockList()) {
-            switch (stockEntity.getSize()) {
-                case S:
-                    stockEntity.setCount(Integer.parseInt(map.get("S").toString()));
-                    break;
-                case M:
-                    stockEntity.setCount(Integer.parseInt(map.get("M").toString()));
-                    break;
-                case L:
-                    stockEntity.setCount(Integer.parseInt(map.get("L").toString()));
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        productRepository.save(productEntity);
-    }
-
     // 24.11.29 - uj
     // 상품 성별 리스트 추출
     @Override
@@ -339,50 +165,4 @@ public class ProductServiceImpl implements ProductService {
         return Arrays.asList(Gender.values());
     }
 
-    // 24.12.11 - 상품 삭제 - sumin
-    @Override
-    public void deleteProduct(int productId) {
-        // productId에 해당하는 상품 조회
-        ProductEntity product = productRepository.findById(productId)
-                .orElseThrow(() -> new RuntimeException("해당 상품이 존재하지 않습니다."));
-        System.out.println("여기와 ?? " + productId);
-        // delete 변수 true로 설정
-        product.setDelete(true);
-
-        // 변경 사항 저장
-        productRepository.save(product);
-
-        System.out.println("상품 삭제 완료 상품 ID: " + productId);
-    }
-
-    // 24.12.11 - 삭제된 상품들 가져오기 - sumin
-    @Override
-    public List<ProductDTO> getDeletedProducts() {
-        List<ProductEntity> deletedProducts = productRepository.findByDeleteTrue(); // 삭제된 상품들 조회
-
-        // Entity -> DTO 변환
-        List<ProductDTO> productList = deletedProducts.stream()
-                .map(productEntity -> ProductDTO.toGetProductDTO(productEntity))
-                .collect(Collectors.toList());
-        System.out.println("productList = " + productList);
-        return productList;
-    }
-
-    @Override
-    public boolean restoreProduct(int productId) {
-        Optional<ProductEntity> productEntity = productRepository.findById(productId);
-
-        if (productEntity.isPresent()) {
-            ProductEntity product = productEntity.get();
-
-            // 이미 삭제 상태인 경우만 복구
-            if (product.getDelete()) {
-                product.setDelete(false);
-                productRepository.save(product);
-                return true; // 복구 성공
-            }
-        }
-
-        return false; // 복구 실패
-    }
 }
